Fix 71a07c7 for 32bits

+++ run.sh
@@ -118,8 +118,12 @@ xattr -w security.selinux u:object_r:system_file:s0 etc/init/media-swcodec.rc
 cp system_ext/apex/com.android.adbd/etc/init.rc etc/init/adbd.rc
 xattr -w security.selinux u:object_r:system_file:s0 etc/init/adbd.rc
 
+if [ -f system_ext/apex/com.android.adbd/lib64/libadb_protos.so ];then
 cp system_ext/apex/com.android.adbd/lib64/libadb_protos.so lib64/libadb_protos.so
 xattr -w security.selinux u:object_r:system_file:s0 lib64/libadb_protos.so
+fi
+cp system_ext/apex/com.android.adbd/lib/libadb_protos.so lib/libadb_protos.so
+xattr -w security.selinux u:object_r:system_file:s0 lib/libadb_protos.so
 
 sed -i s/ro.iorapd.enable=true/ro.iorapd.enable=false/g etc/prop.default
 xattr -w security.selinux u:object_r:system_file:s0 etc/prop.default
